chore: cleaned up PropertyDatabase: ListVariables - ListProperties now is the first chore: fixed error interpolating scalar node variables to element integration points development: added new example that shows how to read selected variables from CSMP binary files misc: fixed typos in screen output

+++ source/includes/run_control/SimulatorSetupParameter.h
@@ -90,7 +90,7 @@ struct SimulatorSetupParameter : Parameter {
     ~SimulatorSetupParameter(){}
 
     VARIABLE_TYPE    type;
-    size_t           vsize;
+    uint32_t         vsize;
     PLACEMENT        placement;
 };
